requirements.txt
lime
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0

pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
joblib>=1.3.0

streamlit>=1.35.0
matplotlib>=3.8.0
requests>=2.32.0
SQLAlchemy>=2.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
email-validator>=2.0.0